+++ Diminuto/fun/luxcontroller.c
@@ -391,7 +391,7 @@ int main(int argc, char ** argv) {
 
             prime = diminuto_controller(&parameters, &state, target, input, output);
 
-            printf("%6d %4llu %6d %6d %6d %6d %6d %6d %6d %6d %6d %6d\n", outputs, elapsed, output, target, input, state.sample, state.proportional, state.integral, state.differential, state.total, state.delta, prime);
+            printf("%6d %4llu %6d %6d %6d %6d %6d %6d %6d %6d %6d %6d\n", outputs, (long long unsigned int)elapsed, output, target, input, state.sample, state.proportional, state.integral, state.differential, state.total, state.delta, prime);
 
             output = prime;
